Add files via upload
Code is stable and has many structural changes, but largely looks the same. I didn't have a chance to merge the code, so it's here instead of in the main branch for now.
I've been frustrated with my efforts to generate a random map. I think I know how to do it now, but that will be another day.